Compiles with egcs 1.1
git-svn-id: https://gforge.info.ucl.ac.be/svn/mozart@7809 ada56829-ad1f-0410-b00f-83cda6628aec

+++ trunk/mozart/platform/emulator/foreign.cc
@@ -1076,7 +1076,7 @@ void term2Buffer(ostream &out, OZ_Term term, int depth)
     smallInt2buffer(out,term);
     break;
   default:
-    out << "<Unknown Tag: " << tag << ">";
+    out << "<Unknown Tag: " << (int) tag << ">";
     break;
   }
 }
